+++ src/BundlerMinifierVsix/Helpers/ProjectHelpers.cs
@@ -102,7 +102,9 @@ namespace BundlerMinifierVsix
             {
                 ProjectItem item = project.ProjectItems.AddFromFile(file);
 
-                if (string.IsNullOrEmpty(itemType) || project.Kind.Equals("{E24C65DC-7377-472B-9ABA-BC803B73C61A}", StringComparison.OrdinalIgnoreCase)) // Website
+                if (string.IsNullOrEmpty(itemType) ||
+                    project.Kind.Equals("{E24C65DC-7377-472B-9ABA-BC803B73C61A}", StringComparison.OrdinalIgnoreCase) || // Website
+                    project.Kind.Equals("{262852C6-CD72-467D-83FE-5EEB1973A190}", StringComparison.OrdinalIgnoreCase))   // Universal apps
                     return;
 
                 item.Properties.Item("ItemType").Value = "None";
@@ -119,13 +121,20 @@ namespace BundlerMinifierVsix
 
             try
             {
-                if (item == null || item.ContainingProject == null || item.ContainingProject.Kind.Equals("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}", StringComparison.OrdinalIgnoreCase))
+                if (item == null ||
+                    item.ContainingProject == null ||
+                    item.ContainingProject.Kind.Equals("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}", StringComparison.OrdinalIgnoreCase)) // ASP.NET 5
                     return;
 
-                if (item.ProjectItems == null) // Website project
+                if (item.ProjectItems == null || // Website project
+                    item.ContainingProject.Kind.Equals("{262852C6-CD72-467D-83FE-5EEB1973A190}", StringComparison.OrdinalIgnoreCase)) // Universal apps
+                {
                     item.ContainingProject.ProjectItems.AddFromFile(newFile);
+                }
                 else
+                {
                     item.ProjectItems.AddFromFile(newFile);
+                }
             }
             catch (Exception ex)
             {